+++ tomcat/src/main/java/nextstep/jwp/controller/Controller.java
@@ -1,7 +1,7 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
 import nextstep.jwp.model.Request;
-import nextstep.jwp.vo.Response;
+import nextstep.jwp.model.Response;
 
 import java.io.IOException;
