+++ tomcat/src/main/java/nextstep/jwp/controller/RequestControllerMapper.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
 import nextstep.jwp.model.Request;
 import nextstep.jwp.vo.RequestMethod;
